Move spi_uefi to library.uefi
